requirement.txt
streamlit
langchain-google-genai
langchain-community
langchain-core
langchain
python-dotenv
beautifulsoup4
reportlab
chromadb
sentence-transformers
pysqlite3-binary